# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StockTank API",
    "version": "0.1.0",
    "description": "Identity, RBAC, public content, advertising, marketing and system endpoints for StockTank. Sessions are cookie-based; state-changing requests must send `X-Requested-With`. Financial content served by this platform is informational only and never investment advice."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin (dev proxy or deployed API host)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "st_session",
        "description": "httpOnly session cookie set by /api/v1/auth/register and /api/v1/auth/login. The server stores only a SHA-256 hash of the token."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "VALIDATION_FAILED",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "INTERNAL"
                ]
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ],
        "description": "Envelope used by every non-2xx response.",
        "example": {
          "error": {
            "code": "VALIDATION_FAILED",
            "message": "Invalid request body",
            "requestId": "7f1c1b7e-2d1a-4a4d-9d2b-3b7b2a1f0c9e",
            "details": {
              "source": "body",
              "issues": [
                {
                  "path": "email",
                  "message": "Invalid email address",
                  "code": "invalid_format"
                }
              ]
            }
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/admin/inquiries": {
      "get": {
        "tags": [
          "Admin: Growth"
        ],
        "summary": "List advertising inquiries",
        "description": "Requires `leads.manage`.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "new",
                "contacted",
                "qualified",
                "proposal_sent",
                "won",
                "lost",
                "spam"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List advertising inquiries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "company": {
                            "type": "string"
                          },
                          "contactName": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "website": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "budgetRange": {
                            "type": "string",
                            "enum": [
                              "under_5k",
                              "from_5k_to_25k",
                              "from_25k_to_100k",
                              "over_100k",
                              "undisclosed"
                            ]
                          },
                          "placementKeys": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "message": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "new",
                              "contacted",
                              "qualified",
                              "proposal_sent",
                              "won",
                              "lost",
                              "spam"
                            ]
                          },
                          "notes": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "advertiserId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "utmSource": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "utmMedium": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "utmCampaign": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "referrer": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "company",
                          "contactName",
                          "email",
                          "website",
                          "budgetRange",
                          "placementKeys",
                          "message",
                          "status",
                          "notes",
                          "advertiserId",
                          "utmSource",
                          "utmMedium",
                          "utmCampaign",
                          "referrer",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "page": {
                      "type": "number"
                    },
                    "pageSize": {
                      "type": "number"
                    },
                    "total": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "items",
                    "page",
                    "pageSize",
                    "total"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "No valid session cookie (UNAUTHENTICATED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Missing X-Requested-With / disallowed Origin, or insufficient permission (FORBIDDEN)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/inquiries/{id}": {
      "patch": {
        "tags": [
          "Admin: Growth"
        ],
        "summary": "Update inquiry",
        "description": "Status, notes and advertiser link. Requires `leads.manage`.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "CSRF guard required on every state-changing request under /api. Any non-empty value (the official clients send \"stocktank\").",
              "example": "stocktank"
            },
            "required": true,
            "description": "CSRF guard required on every state-changing request under /api. Any non-empty value (the official clients send \"stocktank\").",
            "name": "x-requested-with",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "new",
                      "contacted",
                      "qualified",
                      "proposal_sent",
                      "won",
                      "lost",
                      "spam"
                    ]
                  },
                  "notes": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 5000
                  },
                  "advertiserId": {
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update inquiry",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "company": {
                      "type": "string"
                    },
                    "contactName": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "website": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "budgetRange": {
                      "type": "string",
                      "enum": [
                        "under_5k",
                        "from_5k_to_25k",
                        "from_25k_to_100k",
                        "over_100k",
                        "undisclosed"
                      ]
                    },
                    "placementKeys": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "message": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "new",
                        "contacted",
                        "qualified",
                        "proposal_sent",
                        "won",
                        "lost",
                        "spam"
                      ]
                    },
                    "notes": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "advertiserId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "utmSource": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "utmMedium": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "utmCampaign": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "referrer": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "company",
                    "contactName",
                    "email",
                    "website",
                    "budgetRange",
                    "placementKeys",
                    "message",
                    "status",
                    "notes",
                    "advertiserId",
                    "utmSource",
                    "utmMedium",
                    "utmCampaign",
                    "referrer",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "No valid session cookie (UNAUTHENTICATED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Missing X-Requested-With / disallowed Origin, or insufficient permission (FORBIDDEN)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Inquiry not found (NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/subscribers": {
      "get": {
        "tags": [
          "Admin: Growth"
        ],
        "summary": "List newsletter subscribers",
        "description": "Requires `newsletter.manage`.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "confirmed",
                "unsubscribed"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List newsletter subscribers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "pending",
                              "confirmed",
                              "unsubscribed"
                            ]
                          },
                          "source": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "utmSource": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "utmMedium": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "utmCampaign": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "consentAt": {
                            "type": "string"
                          },
                          "confirmedAt": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "unsubscribedAt": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "email",
                          "status",
                          "source",
                          "utmSource",
                          "utmMedium",
                          "utmCampaign",
                          "consentAt",
                          "confirmedAt",
                          "unsubscribedAt",
                          "createdAt"
                        ]
                      }
                    },
                    "page": {
                      "type": "number"
                    },
                    "pageSize": {
                      "type": "number"
                    },
                    "total": {
                      "type": "number"
                    },
                    "counts": {
                      "type": "object",
                      "properties": {
                        "pending": {
                          "type": "integer"
                        },
                        "confirmed": {
                          "type": "integer"
                        },
                        "unsubscribed": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "pending",
                        "confirmed",
                        "unsubscribed"
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "page",
                    "pageSize",
                    "total",
                    "counts"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "No valid session cookie (UNAUTHENTICATED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Missing X-Requested-With / disallowed Origin, or insufficient permission (FORBIDDEN)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/subscribers/export.csv": {
      "get": {
        "tags": [
          "Admin: Growth"
        ],
        "summary": "Export confirmed subscribers (CSV)",
        "description": "Confirmed subscribers only; cells are protected against spreadsheet formula injection. Audited.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "No valid session cookie (UNAUTHENTICATED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Missing X-Requested-With / disallowed Origin, or insufficient permission (FORBIDDEN)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}